+++ _changelog/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add search function by input
 * Cleaned up code 
 * Added comments
+* Add filter by category
 
 # 31/10/2024
 * Add : Model, View & Controller for Admin Role.